Dev: Prune rectangles and batches that are not needed in the OpenGL backend.
If a rectangle is entierly outside of the current clip rect, it will
never be visible and can thus pruned.

If we have just pushed a new clip rect and are now pushing a rectangle,
we don't need to create a new batch if the rectangle is inside both the
current clip and the previous clip rects.