Bugfix(UI): Make the color picker work together with the text fields properly.
This was done by copying the value from the text field to the color if
the field is active. If the field is not active, it copies the color
value into itself for display.

Note that this introduces a frame of delay for the text fields when they
are first shown and whenever the color is changed through the picker.