Dev(UI): Make `ui_get_character_index_from_mouse_pos` unicode aware and change its name to reflect this.
The routine will now parse UTF-8 strings and measure the codepoints.
This currently has no effect on behaviour as no other parts of the
system support unicode yet, but it is one step closer now.

Note that the routine will return indicies that are slightly off if the
font has kerning.